FA-flag: master flag plumbing + GQA-Sum + capacity switch (partial); fix sequence.cpp/sequence.h mismatch left uncommitted in d8fdf45
use_flash_attention becomes a true master flag toggling two internally-coherent attention models (ON=on-chip/FlashAttention, OFF=materialized, new default) across the live GQA path, capacity, and energy counters. This commit covers:

PLUMBING (closes I10, the GQA gap where use_flash_attention was parsed but never reached GQA): attention.h/.cpp add a use_flash_attention member+ctor+Create param to SelfAttentionGen/Sum/Mixed (and MLA-Gen/AbsorbMLAGen/AbsorbMLASum for a uniform master flag), setting layer_info.use_flash_attention in each forward(). parallel.h/.cpp thread it through SelfAttentionParallel and AbsorbMLAParallel into both Gen+Sum Create calls. layer.cpp passes scheduler->system_config.use_flash_attention at both attention construction sites.

GQA-SUM-GPU (attention_sum_impl.cpp): Scoring/Context score+P terms dropped under ON; softmax memory charge added under OFF (2*m*n*heads, closing the gap where GQA's softmax was previously always free) at the correct scarce-tier bandwidth (hbm_read_bandwidth/logic_sram_bandwidth on HBF-family, memory_bandwidth otherwise -- NOT the raw memory_bandwidth variable, which is repurposed to mean flash bandwidth on HBF devices); the ideal/ramulator score I/O energy counters gated behind !use_flash_attention (closes I14).

CAPACITY SWITCH (footprint.h, applied in LOCK-STEP to both cluster.cpp:176 and parallelism_optimizer.cpp:372 so the live-sim and optimizer capacity gates never drift): OFF promotes scoreInclusiveIntermediateBytes from diagnostic-only to the live gate (full resident score, matching CB1's reference behavior); ON keeps the existing 256-col tile. intermediateExtrasBytes gained a use_flash_attention param to skip the tile term under OFF (avoiding double-counting against the now-resident score).

STILL OPEN (separate follow-up commit, non-blocking for paper-1 -- no MLA model in the sweep): GQA-Gen-GPU's own score/softmax gating (attention_gen_impl.cpp) is done and verified but held back from this commit because a background task is still actively completing the MLA/Absorb else-branch gating (ruling #3) and Mixed-family parity (ruling #5) in that same file plus attention_mixed_impl.cpp -- committing mid-edit risked a torn snapshot. Will follow in a second commit once confirmed stable.

sequence.cpp: BUGS #30's int64_t widen (committed in d8fdf45) only touched sequence.h's declarations -- the out-of-line definitions in sequence.cpp still returned int, an unnoticed staging gap that left main with a header/definition type mismatch (would fail to compile from a clean clone). Fixed to match.

+++ src/scheduler/sequence.cpp
@@ -107,8 +107,10 @@ int BatchedSequence::get_process_token() {
   return token;
 }
 
-int BatchedSequence::get_gen_process_token() {
-  int token = 0;
+int64_t BatchedSequence::get_gen_process_token() {
+  // BUGS #30 (CB1): int64_t accumulator -- see sequence.h's comment above
+  // the declaration.
+  int64_t token = 0;
   std::vector<Sequence::Ptr> gen_seq = get_gen();
 
   for (auto& seq : gen_seq) {
@@ -117,8 +119,8 @@ int BatchedSequence::get_gen_process_token() {
   return token;
 }
 
-int BatchedSequence::get_sum_process_token() {
-  int token = 0;
+int64_t BatchedSequence::get_sum_process_token() {
+  int64_t token = 0;
   std::vector<Sequence::Ptr> sum_seq = get_sum();
 
   for (auto& seq : sum_seq) {
@@ -141,9 +143,11 @@ int BatchedSequence::get_average_sequence_length() {
   return total_length;
 }
 
-int BatchedSequence::get_total_sequence_length() {
+int64_t BatchedSequence::get_total_sequence_length() {
+  // BUGS #30 (CB1): int64_t accumulator -- see sequence.h's comment above
+  // the declaration.
   std::vector<Sequence::Ptr> gen_seq = get_gen();
-  int total_length = 0;
+  int64_t total_length = 0;
   for (auto seq : gen_seq) {
     total_length += seq->current_len;
   }
